HW Loops - HW all small fixes

+++ Prep_Course/03_Loops/HomeWork/Problem 6. Word or Number/Problem_06.cs
@@ -36,7 +36,7 @@ namespace Problem_6._Word_or_Number
             }
             else
             {
-                for (int i = (int)input.Length; i > 0 ; i--)
+                for (int i = input.Length; i > 0 ; i--)
                 {
                     result += input[i-1];
                 }
